Feature/auth

+++ src/health-and-fitness/api/routes/user.js
@@ -1,18 +1,24 @@
 import userController from "../controllers/userController.js";
 
 export default async function handler(req, res) {
-  switch (req.method) {
-    case "GET":
-      return userController.getUser(req, res);
-    case "POST":
-      return userController.createUser(req, res);
-    case "PUT":
-      return userController.updateUser(req, res);
-    case "DELETE":
-      return userController.deleteUser(req, res);
-    default:
-      res.setHeader("Allow", ["GET", "POST", "PUT", "DELETE"]);
-      return res.status(405).end(`Method ${req.method} Not Allowed`);
+  if ((req.url.startsWith("/api/user/sign-in")) && req.method === "POST") {
+    return userController.signIn(req, res);
+  }
+
+  if (req.url.startsWith("/api/user")) {
+    switch (req.method) {
+      case "GET":
+        return userController.getUser(req, res);
+      case "POST":
+        return userController.createUser(req, res);
+      case "PUT":
+        return userController.updateUser(req, res);
+      case "DELETE":
+        return userController.deleteUser(req, res);
+      default:
+        res.setHeader("Allow", ["GET", "POST", "PUT", "DELETE"]);
+        return res.status(405).end(`Method ${req.method} Not Allowed`);
+    }
   }
 }
